- [BE] refactor(world-housing): 목표지역 seed 적재 및 provider 구현/property id 형변환 (S14P21C103-240)

See merge request s14-fintech-finance-sub1/S14P21C103!93

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/service/world/result/GameWorldResult.java
@@ -178,14 +178,14 @@ public class GameWorldResult {
     public static class HousingSnapshot {
 
         private final HousingType currentHousingType;
-        private final Long currentPropertyId;
-        private final Long targetPropertyId;
+        private final Integer currentPropertyId;
+        private final Integer targetPropertyId;
         private final boolean hasHousingLossSignal;
 
         private HousingSnapshot(
             HousingType currentHousingType,
-            Long currentPropertyId,
-            Long targetPropertyId,
+            Integer currentPropertyId,
+            Integer targetPropertyId,
             boolean hasHousingLossSignal
         ) {
             this.currentHousingType = currentHousingType;
@@ -196,8 +196,8 @@ public class GameWorldResult {
 
         public static HousingSnapshot of(
             HousingType currentHousingType,
-            Long currentPropertyId,
-            Long targetPropertyId,
+            Integer currentPropertyId,
+            Integer targetPropertyId,
             boolean hasHousingLossSignal
         ) {
             return new HousingSnapshot(
