+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/world/housing/GameContractReviewRepository.java
@@ -7,6 +7,6 @@ public interface GameContractReviewRepository extends JpaRepository<GameContract
 
     Optional<GameContractReview> findTopByGameSessionIdAndPropertyIdOrderByReviewedAtDesc(
         Integer gameSessionId,
-        Long propertyId
+        Integer propertyId
     );
 }
